CVSL-1482 move PolicyVersion enum
